collapse the hierarcy a bit
